Improve shutdown sequence
Let the DefWndProc handle the WM_CLOSE message and post the quit message
in a WM_DESTROY case instead, this is how it is traditionally done so it
may be more stable.

Also use PostMessage to send WM_CLOSE which adds the message to the
queue rather than invoking WndProc immediately. This however can cause
GetMessage to return an error code after the window has been destroyed
so it was necessary to add a boolean so we can ignore such an error
since we are quitting the loop anyway.